Change connection.c:gaim_connection_error() to still work even if the message parameter is NULL.  There's no reason to error-out of that function just because some yuppie didn't pass us an error message. But we still log a critical message, so those of you at home running with fatal_criticals will still crash.
I noticed this because Jabber isn't handling <stream:error/> correctly
right now.  Anyone else notice this?  The code is written to look for
"stream:error", but our xmlnode stuff just sees "error" (even though
the "stream:error" format is sent over the wire).  I think it's
because of the libxml change.

To reproduce the problem, sign onto the same resource twice and note
that the error message that Gaim gives you sucks.