Heres the flow I'm trying to model out in api/Signal.cs
1. a .Net plugin hooks into an event, and Signal.connect eventually gets called with the signal name,
and a delegate into the .Net plugin
2. Signal.cs stores the provided delegate in a hash of signal names => vector of delegates. (Not Implemented)
3. Signal.cs calls _connect and pass in its own delegate instead (Currently hardcoded since we only have one signal),
and passes in the string signal name as the user_data
4. When libpurple core fires the signal, Signal.cs's delegate gets called, uses the signal name provided in the
user_data to look into the hash map, converts IntPtrs to Purple.Objects and calls all the real delegates. (Somewhat implemented)

Basically all I'm doing with step 4 is passing in the string signal name and dumping it to the console, but its not
getting passed around, which means I can't then fire off the real delegates.  Hmmm.