Update the PurpleAPI to use .Net 2.0 constructs, woot for Generics!
Also, api/Signal.cs is now abstracted for use with other signals.
Anticipating that when we generate the bindings we'll also auto generate
all the signal marshalling stuff, I moved that code to Signal.Marshall.cs.

Also, I have a few ideas for handling callbacks that expect something other
than an IntPtr (bool, int, etc).

The next step now is to generate the mappings for all the signals, and then
start generating wrappings for individual modules.