applied changes from bbcc1925d9e449bb9599d378da2764bf47da2a7a              through c2cc31f23a71f23062555721f8101fc9c997bae2
I commited this yesterday, but it includes some other changes that I
didn't mean to commit, so I reverted it and now I'm just committing
this file.

Here's the original commit message:
When we get a presence of type="error" from the server that
corresponds to a Jabber chat, only destroy/free/close/leave
the chat room if the error happened while joining (and we
were therefore never in the room).

This fixes the following bug:
1. Join a room
2. Try to change your nickname to something that's being
   used by someone else
3. The server gives you an error message, but you're not 
   actually kicked out of the room
4. Pidgin thinks you've been kicked out and won't let you 
   send messages to the room, etc.