Very first revision for mono-generate-signals.
This script is responsible for looking a module .c file and picking out purple_signal_register
calls.  

Ideally this script would be way cleaner: using pyparser and a more rigid grammar to parse
out what we need.

Instead I'm using regexes and making some assumptions, but it seems to work.  In short it eats
babies :-)