Use the correct xmlnode when processing SOAP Faults.
When the MSN SOAP server sends a "psf:Redirect" Fault, look in the 
"Fault" node for the "redirectUrl", and not in the (possibly) 
non-existent "Body" node. This should fix login for @msn.com addresses, 
who seem to use a different login server.

When we get a "wsse:FailedAuthentication" Fault, look in the "Fault" 
node for the "faultstring", instead of the "Body" node. Thanks, 
Dimmuxx.